feat: split erf/erfc + USE_PARITY template param + parity x²-Horner
Split ckernel_sfpu_erf_erfc.h into ckernel_sfpu_erf.h (with USE_PARITY=true)
and ckernel_sfpu_erfc.h (standard eval). Added USE_PARITY template parameter
to piecewise_rational_eval so parity can be controlled per-call without macros.

erf BF16: MaxULP 1.0, 23µs Tracy (1.7× faster than original 39µs)
erf FP32: MaxULP 5.9, 38µs Tracy
erfc BF16: MaxULP 126, 42µs Tracy
erfc FP32: MaxULP 21.6M, 46µs Tracy
